Add init command
