+++ client/lib/.config.json
@@ -1,0 +1,1 @@
+{"current_remote":"https://agile-cli.herokuapp.com","current_user":null}
